requirements.txt
anyio==3.7.0
asgiref==3.7.2
certifi==2023.5.7
Django==4.2.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
h11==0.14.0
httpcore==0.17.2
httpx==0.24.1
idna==3.4
PyJWT==2.7.0
python-dotenv==1.0.0
python-telegram-bot==20.3
pytz==2023.3
sniffio==1.3.0
sqlparse==0.4.4
